feat: support tabbed windows
Usually tabbed windows falls short with window management tools, like
Aerospace for example, you need to give up on using Terminal.app or even
Finder.

This commit adds a still experimental and hacky way to deal with it by
handling tabbedWindows differently and keeping track of them, so we can
manipulate them accordingly.

This fixes:

1. focus jumping to another window when a tab is closed, focus now
   remains at the window in which a tab was closed
2. same window present in multiple virtual spaces, it could happen each
   different tab was assigned to different virtual spaces, making the window
   present on both

About this particularity and how other programs deal with it:
- https://github.com/koekeishiya/yabai/issues/68
- https://github.com/nikitabobko/AeroSpace/issues/68